+++ Tests/RubyGatewayTests/TestVM.swift
@@ -177,6 +177,23 @@ class TestVM: XCTestCase {
         XCTAssertEqual(testTitle, Ruby.scriptName)
     }
 
+    /// $SAFE
+    func testSafeMode() {
+        XCTAssertFalse(Ruby.taintChecks)
+        XCTAssertEqual(0, Int(try Ruby.getGlobalVar("$SAFE")))
+        Ruby.taintChecks = true
+        XCTAssertTrue(Ruby.taintChecks)
+        XCTAssertEqual(1, Int(try Ruby.getGlobalVar("$SAFE")))
+        Ruby.taintChecks = false
+        let (mj, mn, _) = Ruby.apiVersion
+        if mj > 2 || mn > 5 {
+            XCTAssertFalse(Ruby.taintChecks)
+            XCTAssertEqual(0, Int(try Ruby.getGlobalVar("$SAFE")))
+        } else {
+            XCTAssertTrue(Ruby.taintChecks)
+        }
+    }
+
     /// Version
     func testVersion() {
         let version = Ruby.version
@@ -222,6 +239,11 @@ class TestVM: XCTestCase {
         XCTAssertFalse(Ruby.debug)
         Ruby.debug = true // swallowed
         XCTAssertFalse(Ruby.debug)
+
+        // SAFE fail-safe
+        XCTAssertFalse(Ruby.taintChecks)
+        Ruby.taintChecks = true // swallowed
+        XCTAssertFalse(Ruby.taintChecks)
 
         // type construction fails out safely
         let strObj: RbObject = "test"